Removed a lot of warnings.

+++ src/ringbuffer.c
@@ -37,7 +37,7 @@
 jack_ringbuffer_t *
 jack_ringbuffer_create (size_t sz)
 {
-  int power_of_two;
+  unsigned int power_of_two;
   jack_ringbuffer_t *rb;
 
   if ((rb = malloc (sizeof (jack_ringbuffer_t))) == NULL)
@@ -45,7 +45,7 @@ jack_ringbuffer_create (size_t sz)
       return NULL;
     }
 
-  for (power_of_two = 1; 1 << power_of_two < sz; power_of_two++);
+  for (power_of_two = 1; (unsigned int) 1 << power_of_two < sz; power_of_two++);
 
   rb->size = 1 << power_of_two;
   rb->size_mask = rb->size;
